HWKMETRICS-469 Add missing ISPN config at hawkular-component level Moving InsertedDataSubscriber into hawkular-component to inject Configuration properties
