+++ hawkular-component/src/test/java/org/hawkular/metrics/api/jaxrs/config/ConfigurableProducer.java
