+++ hawkular-component/src/test/resources/arquillian.xml
