+++ hawkular-component/src/test/resources/jboss-deployment-structure.xml
