+++ hawkular-component/src/test/resources/web.xml
