+++ hawkular-component/src/test/resources/wildfly/configuration/standalone-test.xml
